requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
tensorflow
joblib
numpy
scikit-learn
Pillow
opencv-python